feat(layouts): refactor shared template code and added top navigation bar
There is now no more duplicated layout code (eg. header and footer) between the CMS and WWW codebases; those shared components now live in `views/common/`. This does mean they're coded in EJS and are rendered by Node instead of Angular, which I'm not thrilled about, but for the time being it's the best solution. The good news is that the mobile view especially looks a lot cleaner now, and is more user-friendly thanks to the sticky top navigation bar to help the user get around the site.