chore(grunt): switch from git-changelog to grunt-conventional-changelog
The previously-implemented git-changelog module for automated changelog generating was buggy and seems like it's not being actively maintained, so the bugs are unlikely to be fixed soon. After doing some Googling, I found Brian T. Ford's grunt-conventional-changelog, which seems to be everything I was looking for, and it does it without the bugs (for as far as I can tell so far).