feat(site settings): making changes to the site settings now refreshes them immediately on the backend, rather than requiring a server restart
Wasn't sure whether to classify this as a bug fix or a feature... but let's be optimistic and call it a feature. Before this change, you could change the site settings from the admin dashboard in the CMS, but the Sails server would need to be restarted for the changes to show up. Now, the settings are refreshed on save, so they will be reflected on the site immediately.