feat(Google Analytics): add the ability to turn on Google Analytics tracking to the site settings and admin dashboard
This commit introduces two new fields to the site_settings table, one to control whether Google Analytics tracking should be turned on or off, and the other to store the Google Analytics ID (which will only be used if the former is set to true). All pages will be tracked except for the CMS side, though as of right now there is not yet any special code to account for this being mostly a Single Page App (SPA), so GA may or may not pick up everything. I plan on monitoring the GA tracking on my own test site for a bit before I spend time optimizing the client-side JS for Google Analytics, if at all.