feat(Twitter cards): added a basic implementation of Twitter cards
Again, it's configurable via the site_settings table and the admin dashboard, and off by default. I also had to add a site description field, because a description is required for all cards. The implementation doesn't yet support customized cards for story pages and author profiles because I'm not sure whether that can be done on the frontend, or whether it has to be done on the backend.