feat(Twitter cards): add Twitter cards to individual stories and author profiles
This expands the Twitter card functionality introduced in the previous commit to support customized cards for individual stories and author profiles. This implementation is done entirely on the frontend, which I'm not 100% sure will work correctly, so I'm committing it to the development branch for now so I can test it against Twitter's validator using my own Heroku test site.